Alpha_FV2316
Second version of the game FV2316. The game underwent a reworking of the design as well as the rest and was made from scratch. Many functions are not yet available, such as: “Settings” and “Donation”. A new plot and new textures await you in the update. Author of the game CreepyFaiz